+++ src/app/modules/resource-table/resource-table.component.ts
@@ -35,8 +35,15 @@ import {
 } from '../../shared/fhir-backend/fhir-backend.service';
 import { ResourceTableFilterComponent } from '../resource-table-filter/resource-table-filter.component';
 import { FilterType } from '../../types/filter-type';
-import Resource = fhir.Resource;
 import { CustomDialog } from '../../shared/custom-dialog/custom-dialog.service';
+import Resource = fhir.Resource;
+
+type TableCells = { [key: string]: string };
+
+interface TableRow {
+  cells: TableCells;
+  resource: Resource;
+}
 
 /**
  * Component for loading table of resources
@@ -120,7 +127,7 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
   columns: string[] = [];
   selectedResources = new SelectionModel<Resource>(true, []);
   filtersForm: FormGroup = new FormBuilder().group({});
-  dataSource = new TableVirtualScrollDataSource<Resource>([]);
+  dataSource = new TableVirtualScrollDataSource<TableRow>([]);
   lastResourceElement: HTMLElement;
   // Data is loading
   isLoading$ = new BehaviorSubject(false);
@@ -215,30 +222,16 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
     this.subscription.unsubscribe();
   }
 
-  /**
-   * Sets default columns.
-   * We show default columns on very first run of the application and when
-   * the user doesn't select any columns to show in the column selection dialog.
-   */
-  private setDefaultColumns(): void {
-    const allColumns = this.columnDescriptionsService.getAvailableColumns(
-      this.resourceType,
-      this.context
-    );
-    this.columnDescriptions = allColumns.filter((x) => x.displayByDefault);
-
-    // Save column selections of default
-    this.columnDescriptionsService.setVisibleColumnNames(
-      this.resourceType,
-      this.context,
-      this.columnDescriptions.map((x) => x.element)
-    );
-  }
-
   ngOnChanges(changes: SimpleChanges): void {
     // update resource table if user searches again
     if (changes['resourceStream'] && changes['resourceStream'].currentValue) {
+      const columnsWithData = {};
       this.dataSource.data.length = 0;
+      this.columnDescriptionsService.setColumnsWithData(
+        this.resourceType,
+        this.context,
+        []
+      );
       this.isLoading$.next(true);
       this.liveAnnoncer.announce(
         `The ${this.resourceType} resources loading process has started`
@@ -246,19 +239,39 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
       const startTime = Date.now();
       this.resourceStream.pipe(bufferCount(50)).subscribe(
         (resources) => {
+          const allColumns = this.columnDescriptionsService.getAvailableColumns(
+            this.resourceType,
+            this.context
+          );
+          let columnsWithDataChanged = false;
+          const newRows: TableRow[] = resources.map((resource) => ({
+            resource,
+            cells: allColumns.reduce((desc, columnDesc) => {
+              const cellText = this.getCellStrings(resource, columnDesc).join(
+                '; '
+              );
+              desc[columnDesc.element] = cellText;
+              if (!columnsWithData[columnDesc.element] && cellText) {
+                columnsWithData[columnDesc.element] = true;
+                columnsWithDataChanged = true;
+              }
+              return desc;
+            }, {} as TableCells)
+          }));
+
           if (this.enableClientFiltering) {
             // Move selectable studies to the beginning of table.
-            this.dataSource.data = [...this.dataSource.data, ...resources].sort(
-              (a, b) => {
+            this.dataSource.data = [...this.dataSource.data, ...newRows].sort(
+              (a: TableRow, b: TableRow) => {
                 if (
-                  !this.myStudyIds.includes(a.id) &&
-                  this.myStudyIds.includes(b.id)
+                  !this.myStudyIds.includes(a.resource.id) &&
+                  this.myStudyIds.includes(b.resource.id)
                 ) {
                   return 1;
                 }
                 if (
-                  this.myStudyIds.includes(a.id) &&
-                  !this.myStudyIds.includes(b.id)
+                  this.myStudyIds.includes(a.resource.id) &&
+                  !this.myStudyIds.includes(b.resource.id)
                 ) {
                   return -1;
                 }
@@ -266,7 +279,14 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
               }
             );
           } else {
-            this.dataSource.data = this.dataSource.data.concat(resources);
+            this.dataSource.data = this.dataSource.data.concat(newRows);
+          }
+          if (columnsWithDataChanged) {
+            this.columnDescriptionsService.setColumnsWithData(
+              this.resourceType,
+              this.context,
+              Object.keys(columnsWithData)
+            );
           }
         },
         () => {},
@@ -287,9 +307,7 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
       if (this.enableSelection) {
         this.columns.push('select');
       }
-      if (!this.columnDescriptions.length) {
-        this.setDefaultColumns();
-      }
+
       this.setTableColumns();
     }
   }
@@ -321,10 +339,10 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
           const columnDescription = this.columnDescriptions.find(
             (c) => c.element === key
           );
-          const cellValue = this.getCellStrings(data, columnDescription);
+          const cellValue = data.cells[columnDescription.element];
           const filterType = this.getFilterType(columnDescription);
           if (filterType === FilterType.Autocomplete) {
-            if (!(value as string[]).includes(cellValue.join('; '))) {
+            if (!(value as string[]).includes(cellValue)) {
               return false;
             }
           }
@@ -333,14 +351,12 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
               '\\b' + escapeStringForRegExp(value as string),
               'i'
             );
-            if (!cellValue.some((item) => reCondition.test(item))) {
-              return false;
-            }
+            return reCondition.test(cellValue);
           }
           if (filterType === FilterType.Number) {
             if (
               !ResourceTableComponent.checkNumberFilter(
-                cellValue.join('; '),
+                cellValue,
                 value as string
               )
             ) {
@@ -374,8 +390,8 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
     this.isAllSelected()
       ? this.selectedResources.clear()
       : this.dataSource.data.forEach((row) => {
-          if (this.myStudyIds.includes(row.id)) {
-            this.selectedResources.select(row);
+          if (this.myStudyIds.includes(row.resource.id)) {
+            this.selectedResources.select(row.resource);
           }
         });
   }
@@ -451,13 +467,9 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
       (c) => c.element === sort.active
     );
     const filterType = this.getFilterType(sortingColumnDescription);
-    this.dataSource.data.sort((a: Resource, b: Resource) => {
-      const cellValueA = this.getCellStrings(a, sortingColumnDescription).join(
-        '; '
-      );
-      const cellValueB = this.getCellStrings(b, sortingColumnDescription).join(
-        '; '
-      );
+    this.dataSource.data.sort((a: TableRow, b: TableRow) => {
+      const cellValueA = a.cells[sortingColumnDescription.element];
+      const cellValueB = b.cells[sortingColumnDescription.element];
       return filterType === FilterType.Number
         ? (+cellValueA - +cellValueB) * (isAsc ? -1 : 1)
         : cellValueA.localeCompare(cellValueB) * (isAsc ? -1 : 1);
@@ -474,13 +486,10 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
     const header = columnDescriptions
       .map((columnDescription) => columnDescription.displayName)
       .join(',');
-    const rows = this.dataSource.data.map((resource) =>
+    const rows = this.dataSource.data.map((row) =>
       columnDescriptions
         .map((columnDescription) => {
-          const cellText = this.getCellStrings(
-            resource,
-            columnDescription
-          ).join('; ');
+          const cellText = row.cells[columnDescription.element];
           if (/["\s,]/.test(cellText)) {
             // According to RFC-4180 which describes common format for CSV files:
             // Fields containing line breaks (CRLF), double quotes, and commas
@@ -511,8 +520,10 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
    */
   setSelectedIds(ids: string[]): void {
     this.selectedResources.clear();
-    const items = this.dataSource.data.filter((r) => ids.includes(r.id));
-    this.selectedResources.select(...items);
+    const items = this.dataSource.data.filter((r) =>
+      ids.includes(r.resource.id)
+    );
+    this.selectedResources.select(...items.map((r) => r.resource));
   }
 
   /**
@@ -532,8 +543,8 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
     const filterType = this.getFilterType(column);
     let options: string[] = [];
     if (filterType === FilterType.Autocomplete) {
-      const columnValues = this.dataSource.data.map((row) =>
-        this.getCellStrings(row, column).join('; ')
+      const columnValues = this.dataSource.data.map(
+        (row) => row.cells[column.element]
       );
       options = [...new Set(columnValues)]
         .filter((v) => v)
@@ -571,7 +582,7 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
   getFilterType(column: ColumnDescription): FilterType {
     return this.listFilterColumns.includes(column.element)
       ? FilterType.Autocomplete
-      : column.displayName.startsWith('Number of')
+      : column.types.length === 1 && column.types[0] === 'Count'
       ? FilterType.Number
       : FilterType.Text;
   }